various changes

+++ todo.py
@@ -1,19 +1,22 @@
 #!/usr/bin/env python3
 
 import argparse
-from collections import namedtuple
+from collections import defaultdict, namedtuple
 import configparser
+import datetime
 from enum import Enum
 import os
 import shutil
 
 
+ACTION_STATE_MARKER = '!'
 class action_states(Enum):
   ACTIONABLE='^..^'
-  WAIT='Zzz'
+  WAIT='zzz'
+  PRIORITY='!'
 
 
-class todo(namedtuple('todo', 'text tags action_state is_completed number projects')):
+class todo(namedtuple('todo', 'text tags action_state is_completed number projects create_date')):
   @classmethod
   def fromline(cls, line):
     toks = line.rstrip().split()
@@ -22,6 +25,7 @@ class todo(namedtuple('todo', 'text tags action_state is_completed number projec
     projects = set()
     action_state = action_states.ACTIONABLE
     done = False
+    date = None
     for i, tok in enumerate(toks):
       if i == 0 and tok == 'X':
         done = True
@@ -29,33 +33,49 @@ class todo(namedtuple('todo', 'text tags action_state is_completed number projec
         tags.add(tok[1:])
       elif tok[0] == '@':
         projects.add(tok[1:])
-      elif tok == '!' + action_states.WAIT.value:
+      elif tok == ACTION_STATE_MARKER + action_states.WAIT.value:
         action_state = action_states.WAIT
+      elif tok == ACTION_STATE_MARKER + action_states.PRIORITY.value:
+        action_state = action_states.PRIORITY
       else:
-        text.append(tok)
+        try:
+          date = datetime.datetime.strptime(tok, '%Y-%m-%d').date()
+        except ValueError:
+          text.append(tok)
 
-    return todo(text=' '.join(text), tags=tags, action_state=action_state, is_completed=done, number=-1, projects=projects)
+    return todo(text=' '.join(text), tags=tags, action_state=action_state, is_completed=done, number=-1, projects=projects, create_date=date)
 
-  def rawstr(self):
-    s = ('X ' if self.is_completed else '') + self.text + ' ' + ' '.join('#' + tag for tag in self.tags) + ' '.join('@' + project for project in self.projects)
-    if self.action_state == action_states.WAIT:
+  def rawstr(self, suppress_action_state=False):
+    s = ('X ' if self.is_completed else '') + self.text + ' ' + ' '.join('#' + tag for tag in self.tags) + ' '.join('@' + project for project in self.projects) + (' ' + str(self.create_date) if self.create_date else '')
+    if not suppress_action_state and self.action_state in (action_states.WAIT, action_states.PRIORITY):
       s += ' !' + self.action_state.value
 
     return s
 
 
 def safeopen(flags):
+  if not os.path.exists(filepath):
+     return open(filepath, 'w')
   shutil.copyfile(filepath, filepath + '-bak')
   return open(filepath, flags)
 
 
 def add(args):
-  safeopen('a+').write(' '.join(args.text) + '\n')
+  serialized = ' '.join(args.text) + ' ' + datetime.date.today().isoformat()
+  safeopen('a+').write(serialized + '\n')
+
+
+def add_silent(args):
+  serialized = ' '.join(args.text + [ACTION_STATE_MARKER + action_states.WAIT.value, datetime.date.today().isoformat()])
+  safeopen('a+').write(serialized + '\n')
 
 
 def readtodos(filter_=None):
   if not filter_:
     filter_ = lambda x: x
+
+  if not os.path.exists(filepath):
+      return []
 
   return filter(filter_, (todo.fromline(line)._replace(number=i) for i, line in enumerate(open(filepath))))
 
@@ -73,11 +93,50 @@ def updatetodo(number, func):
   safeopen('w+').write('\n'.join(todo.rawstr() for todo in todos) + '\n')
 
 
-def printtodos(filter_=None):
+def printtodos(filter_, show_action_state, suppress_number):
   todos = list(readtodos(filter_))
+  if not todos:
+      return
   digits = len(str(todos[-1].number))
   for todo in todos:
-    print(str(todo.number).zfill(digits), todo.rawstr())
+      n = str(todo.number).zfill(digits) if not suppress_number else '-'
+      print(n, todo.rawstr(not show_action_state))
+
+
+def lsa(args):
+    args.all = True
+    return ls(args)
+
+
+def lp(args):
+  def filter_(x):
+    return len(x.projects) and not x.is_completed
+
+  todos = list(readtodos(filter_))
+  projects = defaultdict(list)
+  for t in todos:
+      for p in t.projects:
+          projects[p].append(t)
+
+  if not todos:
+      return
+  digits = len(str(todos[-1].number))
+  for i, (p, tod) in enumerate(projects.items()):
+      print('#', p)
+      for todo in tod:
+          n = str(todo.number).zfill(digits)
+          print('-', n, todo.rawstr())
+      if i < len(projects) - 1:
+          print()
+
+completed_filter = lambda x: not x.is_completed
+
+
+def np(args):
+  def filter_(todo):
+      return not todo.projects
+
+  printtodos(lambda x: completed_filter(x) and filter_(x), hasattr(args, 'all') and args.all, hasattr(args, 'export') and args.export)
 
 
 def ls(args):
@@ -92,36 +151,63 @@ def ls(args):
     return set(todo.projects).intersection(projects)
 
   filter_ = lambda x: True
-  completed_filter = lambda x: not x.is_completed
 
   if hasattr(args, 'waiting') and args.waiting:
     filter_ = lambda x: x.action_state == action_states.WAIT
-  elif not hasattr(args, 'all'):
-    filter_ = lambda x: x.action_state == action_states.ACTIONABLE
-  printtodos(lambda x: completed_filter(x) and filter_(x) and filterprojects(x))
+  elif not hasattr(args, 'all') or not args.all:
+    filter_ = lambda x: x.action_state == action_states.PRIORITY
+
+  tags = set(getattr(args, 'tags', []))
+  if tags:
+      filter_ = lambda x: set(x.tags) & tags
+  printtodos(lambda x: completed_filter(x) and filter_(x) and filterprojects(x), hasattr(args, 'all') and args.all, hasattr(args, 'export') and args.export)
+
+
+def update_action_state(action_state):
+    def inner(t):
+        return t._replace(action_state=action_state)
+    return inner
 
 
 def mark_actionable(args):
-  def update(t):
-    return t._replace(action_state=action_states.ACTIONABLE)
-
-  updatetodo(args.number[0], update)
+  update_all(args.number, update_action_state(action_states.ACTIONABLE))
 
 
 def mark_waiting(args):
-  def update(t):
-    return t._replace(action_state=action_states.WAIT)
+  update_all(args.number, update_action_state(action_states.WAIT))
 
-  updatetodo(args.number[0], update)
+
+def mark_priority(args):
+  update_all(args.number, update_action_state(action_states.PRIORITY))
 
 
 def mark_done(args):
   def update(t):
     if t.is_completed:
         raise Exception('Already marked completed.')
+    print('Marked completed: "%s"' % t.text)
     return t._replace(is_completed=True)
 
-  updatetodo(args.number[0], update)
+  return update_all(args.number, update)
+
+
+def parse_numbers(numbers):
+    num = []
+    for n in numbers:
+        if type(n) == int:
+            num.append(n)
+            continue
+        n = n.split('-')
+        assert len(n) <= 2
+        if len(n) == 2:
+            num += list(range(int(n[0]), int(n[1]) + 1))
+        else:
+            num.append(int(n[0]))
+
+    return num
+
+def update_all(numbers, fn):
+  return [updatetodo(ni, fn) for ni in sorted(parse_numbers(numbers), reverse=True)]
 
 
 def add_project(args):
@@ -133,7 +219,11 @@ def add_project(args):
     projects.add(project)
     return t._replace(projects=projects)
 
-  updatetodo(args.number[0], update)
+  number = [arg for arg in args.number if arg[0] != '@']
+  projects = [arg for arg in args.number if arg[0] == '@']
+  args.number = number
+  args.project = projects[0]
+  return update_all(args.number, update)
 
 
 def edit(args):
@@ -156,27 +246,45 @@ if __name__ == '__main__':
   sub_parser.add_argument('text', nargs='+')
   sub_parser.set_defaults(func=add)
 
+  sub_parser = subparsers.add_parser('add-silent', aliases=['as'])
+  sub_parser.add_argument('text', nargs='+')
+  sub_parser.set_defaults(func=add_silent)
+
   sub_parser = subparsers.add_parser('list', aliases=['ls'])
   sub_parser.add_argument('--all', action='store_true')
   sub_parser.add_argument('--waiting', action='store_true')
+  sub_parser.add_argument('--export', action='store_true')
   sub_parser.add_argument('project', nargs='*')
   sub_parser.set_defaults(func=ls)
 
+  sub_parser = subparsers.add_parser('list-all', aliases=['lsa'])
+  sub_parser.add_argument('tags', nargs='*')
+  sub_parser.set_defaults(func=lsa)
+
+  sub_parser = subparsers.add_parser('list-projects', aliases=['lp'])
+  sub_parser.set_defaults(func=lp)
+
+  sub_parser = subparsers.add_parser('no-projects', aliases=['np'])
+  sub_parser.set_defaults(func=np)
+
+  sub_parser = subparsers.add_parser('mark-priority', aliases=['mp'])
+  sub_parser.add_argument('number', nargs='+', type=str)  # TODO
+  sub_parser.set_defaults(func=mark_priority)
+
   sub_parser = subparsers.add_parser('mark-actionable', aliases=['ma'])
-  sub_parser.add_argument('number', nargs='+', type=int)  # TODO
+  sub_parser.add_argument('number', nargs='+', type=str)  # TODO
   sub_parser.set_defaults(func=mark_actionable)
 
   sub_parser = subparsers.add_parser('mark-waiting', aliases=['mw'])
-  sub_parser.add_argument('number', nargs='+', type=int)  # TODO
+  sub_parser.add_argument('number', nargs='+', type=str)  # TODO
   sub_parser.set_defaults(func=mark_waiting)
 
   sub_parser = subparsers.add_parser('do', aliases=['x'])
-  sub_parser.add_argument('number', nargs='+', type=int)  # TODO
+  sub_parser.add_argument('number', nargs='+', type=str)  # TODO
   sub_parser.set_defaults(func=mark_done)
 
-  sub_parser = subparsers.add_parser('proj')
-  sub_parser.add_argument('number', nargs='+', type=int)  # TODO
-  sub_parser.add_argument('project')  # TODO
+  sub_parser = subparsers.add_parser('proj', aliases=['p'])
+  sub_parser.add_argument('number', nargs='+', type=str)  # TODO
   sub_parser.set_defaults(func=add_project)
 
   sub_parser = subparsers.add_parser('edit', aliases=['e'])
